move DoK API behind an interface
Issue: #2
